create hooks useMovieDetails and useMovieCredits
